requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Migrate
python-dotenv
openpyxl
fpdf
Flask-WTF
email-validator
Flask-Bootstrap
wtforms-sqlalchemy
psycopg2-binary
gunicorn
